requirements.txt
ultralytics
opencv-python
requests
fastapi
uvicorn
numpy
pydantic
python-dotenv